feat: footer, cards, ajeitando botoes
