+++ src/features/system-adm/components/ui/confirm-modal.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { X } from 'lucide-react';
+import { Button } from './button'; // Importe seu componente Button
 
 interface ConfirmModalProps {
     message: string;
     buttonmsg: string;
-  onConfirm: () => void;
-  onCancel: () => void;
-  isOpen: boolean;
+    onConfirm: () => void;
+    onCancel: () => void;
+    isOpen: boolean;
 }
 
 export const ConfirmModal: React.FC<ConfirmModalProps> = ({
@@ -15,32 +16,33 @@ export const ConfirmModal: React.FC<ConfirmModalProps> = ({
     onConfirm,
     onCancel,
     isOpen,
-  }) => {
+}) => {
     if (!isOpen) return null;
   
-return (
-    <div className="fixed inset-0 flex items-center justify-center bg-black/20 z-50">
-    <div className="bg-white p-8 rounded shadow-lg text-black w-[420px] relative">
-        <button
-        onClick={onCancel}
-        className="absolute top-2 right-2 text-gray-500 hover:text-black"
-        >
-        <X className="w-5 h-5" />
-        </button>
+    return (
+        <div className="fixed inset-0 flex items-center justify-center bg-black/20 z-50">
+            <div className="bg-white p-8 rounded shadow-lg text-black w-[420px] relative">
+                <button
+                    onClick={onCancel}
+                    className="absolute top-2 right-2 text-gray-500 hover:text-black"
+                >
+                    <X className="w-5 h-5" />
+                </button>
 
-        <p className="mb-6 text-lg font-semibold">
-        {message}
-        </p>
+                <p className="mb-6 text-lg font-semibold">
+                    {message}
+                </p>
 
-        <div className="flex justify-end">
-        <button
-            onClick={onConfirm}
-            className="px-3 py-1.5 bg-red-600 hover:bg-red-700 text-white rounded"
-        >
-            {buttonmsg}
-        </button>
+                <div className="flex justify-end">
+                    <Button
+                        onClick={onConfirm}
+                        type="button"
+                        className="!bg-red-600 hover:!bg-red-700 !px-3 !py-1.5 !h-auto !w-auto !text-base"
+                    >
+                        {buttonmsg}
+                    </Button>
+                </div>
+            </div>
         </div>
-    </div>
-    </div>
     );
 };
